Commit early project history
This commit is to update the latest commit message to one in keeping
with commit message best practices; there are no file changes.  The
subject line in the message of the prior commit was very long and
messed up the output of commands like "git branch -avv"; this commit
is the fix.